Rename fakeNavigation.ts to fakeSceneProps.ts
This file's scope is going to expand.
