+++ src/__tests__/components/BuyCrypto.test.tsx
@@ -4,8 +4,8 @@ import renderer from 'react-test-renderer'
 
 import { BuyCrypto } from '../../components/themed/BuyCrypto'
 import { initialState } from '../../reducers/scenes/SettingsReducer'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
 import { FakeProviders, FakeState } from '../../util/fake/FakeProviders'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('BuyCrypto', () => {
   const mockState: FakeState = {
