+++ src/__tests__/components/SelectFioAddress.test.tsx
@@ -4,7 +4,7 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { getTheme } from '../../components/services/ThemeContext'
 import { SelectFioAddressComponent } from '../../components/themed/SelectFioAddress'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('SelectFioAddress', () => {
   it('should render with loading props', () => {
