+++ src/__tests__/scenes/CreateWalletSelectFiatScene.test.tsx
@@ -4,9 +4,9 @@ import TestRenderer from 'react-test-renderer'
 
 import { CreateWalletSelectFiatScene } from '../../components/scenes/CreateWalletSelectFiatScene'
 import { RouteProp } from '../../types/routerTypes'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
 import { fakeNonce } from '../../util/fake/fakeNonce'
 import { FakeProviders, FakeState } from '../../util/fake/FakeProviders'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('CreateWalletSelectFiatComponent', () => {
   const nonce = fakeNonce(0)
