+++ src/__tests__/scenes/FioAddressDetailsScene.test.tsx
@@ -4,8 +4,8 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { FioAddressDetails } from '../../components/scenes/Fio/FioAddressDetailsScene'
 import { getTheme } from '../../components/services/ThemeContext'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
 import { fakeNonce } from '../../util/fake/fakeNonce'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('FioAddressDetails', () => {
   const nonce = fakeNonce(0)
