+++ src/__tests__/scenes/FioAddressRegisterScene.test.tsx
@@ -4,7 +4,7 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { FioAddressRegister } from '../../components/scenes/Fio/FioAddressRegisterScene'
 import { getTheme } from '../../components/services/ThemeContext'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('FioAddressRegister', () => {
   it('should render with loading props', () => {
