+++ src/__tests__/scenes/FioAddressRegisteredScene.test.tsx
@@ -4,8 +4,8 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { FioAddressRegistered } from '../../components/scenes/Fio/FioAddressRegisteredScene'
 import { getTheme } from '../../components/services/ThemeContext'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
 import { fakeNonce } from '../../util/fake/fakeNonce'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('FioAddressRegistered', () => {
   const nonce = fakeNonce(0)
