+++ src/__tests__/scenes/FioDomainRegisterScene.test.tsx
@@ -4,7 +4,7 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { FioDomainRegister } from '../../components/scenes/Fio/FioDomainRegisterScene'
 import { getTheme } from '../../components/services/ThemeContext'
-import { fakeNavigation } from '../../util/fake/fakeNavigation'
+import { fakeNavigation } from '../../util/fake/fakeSceneProps'
 
 describe('FioDomainRegister', () => {
   it('should render with loading props', () => {
